refactor(beast-interpreter): split parameter_map into submodules with CompiledExpr newtype
Issue #76: parameter_map.rs was a ~408-LOC multi-responsibility hotspot
owning AST definition, parsing, symbol resolution, evaluation, reference
extraction, and literal parsing in a single module. Split into a module
directory with focused sub-modules:

  parameter_map/
  |-- mod.rs          facade + re-exports
  |-- ast.rs          ExprNode (crate-private) + CompiledExpr newtype
  |-- parser.rs       recursive-descent grammar rules
  |-- literal.rs      Q32.32 dot-decimal literal parsing
  |-- eval.rs         pure saturating-arithmetic evaluator
  |-- analysis.rs     collect_channel_refs static pass
  |-- test_support.rs shared channel-manifest fixtures (cfg(test))

The AST enum's variants are now pub(crate), so only the parser can mint
a CompiledExpr -- "channel symbols resolved at parse time" is enforced in
the type system, not by convention. Expr is kept as a type alias for
CompiledExpr so downstream call sites (composition.rs, emission.rs,
tests/common/mod.rs) compile unchanged.

Pure refactor: zero behavioural change. All 88 unit tests + integration
tests + doc tests pass (previously 84 -- the extra 4 cover the new
literal helpers). Keeps the existing recursive-descent logic byte-for-byte;
no new operators (that is issue #61).
