+++ crates/beast-interpreter/src/lib.rs
@@ -59,6 +59,8 @@ pub use emission::{emit_primitives, ACTIVATION_COST_PARAM};
 pub use error::{InterpreterError, Result};
 pub use expression::filter_hooks_by_affordances;
 pub use interpreter::interpret_phenotype;
-pub use parameter_map::{collect_channel_refs, eval_expression, parse_expression, Expr};
+pub use parameter_map::{
+    collect_channel_refs, eval_expression, parse_expression, CompiledExpr, Expr,
+};
 pub use phenotype::{BodyRegion, BodySite, Environment, LifeStage, ResolvedPhenotype};
 pub use scale_band::apply_scale_band_filter;
